Fixed endgame, non-recursive launcher

+++ src/endgame.h
@@ -7,10 +7,10 @@
 #include "map.h"
 
 typedef struct _score {
-	size_t count;
-	size_t stuck;
-	size_t command;
-	size_t total;
+	int count;
+	int stuck;
+	int command;
+	int total;
 } Score;
 
 typedef struct _endscreen {
@@ -23,7 +23,7 @@ typedef struct _endscreen {
 int checkEndGame(Textures *gameTextures, Tile **mapTiles);
 void countTanks(Score *score, Tile **mapTiles, int tank, int base, int origin);
 void countPoints(Tile **mapTiles, Score *score, int tank, int origin);
-int endGame(Textures *gameTextures, size_t score1, size_t score2);
+int endGame(Textures *gameTextures, int score1, int score2);
 void loadScreen(EndScreen *endscreen, int score1, int score2);
 
 #endif
